Refactored NodeOptions type from node.*.ts, updated tests
