+++ types/node.d.ts
@@ -146,44 +146,6 @@ declare module 'rclnodejs' {
   }
 
 
-/**
- * NodeOptions specify configuration choices during the 
- * node instantiation process.  
- * @class
- */
-class NodeOptions {
-
-  /**
-   * Create a new instance with default property values.
-   */
-  constructor();
-
-  /**
-   * When true  
-   * Default value = true;
-   * @returns {boolean} - 
-   */
-  startParameterServices: boolean;
-
-  /**
-   * An array of Parameters that serve as overrides for a node's default
-   * parameters. Default = empty array [].
-   */
-  parameterOverrides: Array<Parameters.Parameter>;
-
-  /**
-   * True indicates that a node shold declare declare parameters from
-   * it's parameter-overrides 
-   */
-  automaticallyDeclareParametersFromOverrides: boolean;
-  
-
-  /**
-   * Returns an instance configured with default options.
-   */
-  static defaultOptions: NodeOptions; 
-} 
-
 	/**
 	 * Node is the primary entrypoint in a ROS system for communication.
 	 * It can be used to create ROS entities such as publishers, subscribers,
